refactor(translation): split the google seam send/forks into translation_seam_google_send (600-line cap)
